remake function have accomplished

+++ src/main/java/com/hfnu/study/community/dto/QuestionDTO.java
@@ -5,12 +5,12 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private long gmtCreate;
     private long gmtModified;
-    private Integer createor;
+    private Long createor;
     private Integer commentAccount;
     private Integer viewAccount;
     private Integer likeCount;
